requirements.txt
aioboto3==10.1.0
aiobotocore==2.4.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aioitertools==0.11.0
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.2
anyio==3.7.1
APScheduler==3.10.1
alembic==1.13.2
async-stripe==4.2.0
async-timeout==4.0.3
asyncer==0.0.2
asyncpg==0.27.0
attrs==24.2.0
boto3==1.24.59
botocore==1.27.59
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==40.0.1
ecdsa==0.19.0
fastapi==0.85.0
fastapi-async-sqlalchemy==0.3.12
fastapi-cache2==0.2.1
frozenlist==1.3.3
gmqtt==0.6.16
greenlet==2.0.2
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
iniconfig==2.0.0
isort==5.10.1
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
mypy==0.982
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pendulum==3.0.0
pillow==10.4.0
platformdirs==3.5.1
pluggy==1.0.0
psycopg==3.1.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.22
pydantic==1.10.17
PyJWT==2.7.0
pypng==0.20220715.0
pytest==7.2.1
pytest-asyncio==0.23.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
pytz==2023.3
PyYAML==5.3.1
qrcode==7.4.2
redis==4.6.0
reportlab==4.0.5
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.1
sendgrid==6.10.0
sentry-sdk==1.9.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starkbank-ecdsa==2.2.0
starlette==0.20.4
stripe==4.2.0
time-machine==2.15.0
tornado==6.4.1
twilio==7.16.4
typing_extensions==4.6.3
tzdata==2024.1
tzlocal==5.0.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
pytest-asyncio==0.23.3
gmqtt==0.6.16
opensearch-py>=2.0.0
